Tidied up actions and actionTypes.
