+++ src/App.css
@@ -2,6 +2,10 @@
   box-sizing: border-box;
 }
 
+body {
+  margin: 0;
+}
+
 .canvas-container {
   height: 600px;
   overflow: auto;
