+++ src/actionTypes/channelActionTypes.js
@@ -1,0 +1,4 @@
+export const ADD_CHANNEL = 'ADD_CHANNEL';
+export const REMOVE_CHANNEL = 'REMOVE_CHANNEL';
+export const UPDATE_CHANNEL_NAME = 'UPDATE_CHANNEL_NAME';
+export const UPDATE_CHANNEL_COLOR = 'UPDATE_CHANNEL_COLOR';
