+++ src/actionTypes/effectActionTypes.js
@@ -1,0 +1,4 @@
+export const ADD_EFFECT = 'ADD_EFFECT';
+export const REMOVE_EFFECT = 'REMOVE_EFFECT';
+export const UPDATE_EFFECT_SETTINGS = 'UPDATE_EFFECT_SETTINGS';
+export const UPDATE_ONE_EFFECT_SETTING = 'UPDATE_ONE_EFFECT_SETTING';
