+++ src/actionTypes/index.js
@@ -1,38 +1,6 @@
-export const ADD_NOTE = 'ADD_NOTE';
-export const REMOVE_NOTE = 'REMOVE_NOTE';
-
-export const ADD_NOTES = 'ADD_NOTES';
-export const REMOVE_NOTES = 'REMOVE_NOTES';
-
-export const PLAY_TRACK = 'PLAY_TRACK';
-export const STOP_TRACK = 'STOP_TRACK';
-
-export const MUTE_MASTER = 'MUTE_MASTER';
-export const UNMUTE_MASTER = 'UNMUTE_MASTER';
-
-export const SET_MASTER_VOLUME = 'SET_MASTER_VOLUME'; 
-
-export const SET_BPM = 'SET_BPM';
-
-export const ADD_CHANNEL = 'ADD_CHANNEL';
-export const REMOVE_CHANNEL = 'REMOVE_CHANNEL';
-
-export const UPDATE_CHANNEL_NAME = 'UPDATE_CHANNEL_NAME';
-export const UPDATE_CHANNEL_COLOR = 'UPDATE_CHANNEL_COLOR';
-
-export const ADD_SECTION = 'ADD_SECTION';
-export const REMOVE_SECTION = 'REMOVE_SECTION';
-
-export const ADD_INSTRUMENT = 'ADD_INSTRUMENT';
-export const REMOVE_INSTRUMENT = 'REMOVE_INSTRUMENT';
-
-export const ADD_EFFECT = 'ADD_EFFECT';
-export const REMOVE_EFFECT = 'REMOVE_EFFECT';
-
-export const UPDATE_INSTRUMENT_SETTINGS = 'UPDATE_INSTRUMENT_SETTINGS';
-export const UPDATE_ONE_INSTRUMENT_SETTING = 'UPDATE_ONE_INSTRUMENT_SETTING';
-export const UPDATE_EFFECT_SETTINGS = 'UPDATE_EFFECT_SETTINGS';
-export const UPDATE_ONE_EFFECT_SETTING = 'UPDATE_ONE_EFFECT_SETTING';
-
-export const OPEN_WINDOW = 'OPEN_WINDOW';
-export const CLOSE_WINDOW = 'CLOSE_WINDOW';
+export * from './sectionActionTypes';
+export * from './playerActionTypes';
+export * from './channelActionTypes';
+export * from './instrumentActionTypes';
+export * from './effectActionTypes';
+export *  from './windowActionTypes';
