+++ src/actionTypes/instrumentActionTypes.js
@@ -1,0 +1,4 @@
+export const ADD_INSTRUMENT = 'ADD_INSTRUMENT';
+export const REMOVE_INSTRUMENT = 'REMOVE_INSTRUMENT';
+export const UPDATE_INSTRUMENT_SETTINGS = 'UPDATE_INSTRUMENT_SETTINGS';
+export const UPDATE_ONE_INSTRUMENT_SETTING = 'UPDATE_ONE_INSTRUMENT_SETTING';
