+++ src/actionTypes/playerActionTypes.js
@@ -1,0 +1,6 @@
+export const PLAY_TRACK = 'PLAY_TRACK';
+export const STOP_TRACK = 'STOP_TRACK';
+export const MUTE_MASTER = 'MUTE_MASTER';
+export const UNMUTE_MASTER = 'UNMUTE_MASTER';
+export const SET_MASTER_VOLUME = 'SET_MASTER_VOLUME'; 
+export const SET_BPM = 'SET_BPM';
