+++ src/actionTypes/sectionActionTypes.js
@@ -1,0 +1,6 @@
+export const ADD_SECTION = 'ADD_SECTION';
+export const REMOVE_SECTION = 'REMOVE_SECTION';
+export const ADD_NOTE = 'ADD_NOTE';
+export const REMOVE_NOTE = 'REMOVE_NOTE';
+export const ADD_NOTES = 'ADD_NOTES';
+export const REMOVE_NOTES = 'REMOVE_NOTES';
