+++ src/actionTypes/windowActionTypes.js
@@ -1,0 +1,2 @@
+export const OPEN_WINDOW = 'OPEN_WINDOW';
+export const CLOSE_WINDOW = 'CLOSE_WINDOW';
