+++ src/index.js
@@ -1,16 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import { Provider } from 'react-redux';
 import store from './store';
-//import Section from './components/Section';
-
-
-//const section = new Section();
-//window.s = section;
-//export const sectionRef = section;
 
 ReactDOM.render(
     <Provider store={store}>
