+++ src/reducers/index.js
@@ -1,5 +1,4 @@
 import { combineReducers } from 'redux';
-import sectionInfo from './sectionInfoReducer';
 import playerInfo from './playerInfoReducer';
 import channels from './channelsReducer';
 import activeWindows from './activeWindowsReducer';
@@ -8,7 +7,6 @@ import instruments from './instrumentsReducer';
 import effects from './effectsReducer';
 
 export default combineReducers({
-    sectionInfo,
     playerInfo,
     activeWindows,
     channels,
